Updated mediawiki/extensions Project: mediawiki/extensions/NavigationTiming  ca423fc89ee549f5ddc62c5b0635bebff64dc4ef
Ensure timings are reported after the page has loaded

We schedule the reporting of Navigation Timing data to occur after the document
has loaded by binding a handler to $( window ).load. This is not reliable,
because the load event may have already occurred at the time we register our
handler, in which case the callback will never fire.

To correct this, bind a handler for the resourceloader.loadEnd hook which
starts an interval timer that checks performance.timing.loadEventEnd. If its
value is nonzero, or if the Navigation Timing API is not available at all, then
the timer cancels itself and calls the callback.

Change-Id: I2605c746b1749b7e48edec89eded4b253bdade23