Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  db2ac21e753841b0a3d412297c3b6e813ac7f308
Enable communication with multiple datacenters

This patch removes $wgCirrusSearchServers and replaces it with two
new variables, $wgCirrusSearchDefaultCluster and $wgCirrusSearchClusters.
Additionally this groups together all per-cluster configuration
under the $wgCirrusSearchClusters key. In the first pass this is just
for timeouts. Another new global, $wgCirrusSearchClusterDefaults
specifys the default values of these configuration values and they
can be overrided by per-cluster configuration.

I'm not a huge fan of the existing code to be backwards compatibile
with $wgCirrusSeachServers.  I plan to strip all of that back out
after we deploy the updated configuration in production.

Almost everything goes though $wgCirrusSearchDefaultCluster. The only
exception is the ElasticaWrite job. This job runs for each cluster.
If any cluster fails an operation a job for just that cluster is
inserted to the job queue.

All jobs now accept a 'cluster' argument. All maintenance scripts
also accept a 'cluster' argument. Both will write to all clusters
when no cluster is specified, and only the specified cluster when
provided.

Bug: T86781
Change-Id: Ie6a559249b863f1e6c1390ec2bb5ed83d2d1ecb8