Updated mediawiki/extensions Project: mediawiki/extensions/TemplateData  fdd051402417431cd73199383e9711e4eaa7619f
Go back to using a configurable link for documentation

The message allows per-wiki configuration. The default target is better
for normal users of TemplateData than the previous hard-coded one and
has already been translated in some languages from when it was used
previously.

This also moves the messages dependencies from `ext.templateData` to
`ext.templateDataGenerator.ui`, which is the only module that uses
them: the correct strings were displayed only by inheritance.

Bug: T112011
Change-Id: I81ab684bb0ed9b0767df75635284c51129ac7000