Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaEvents  2662daae4fdeef8e3b4682ea6822265f57725352
Merge ext.wikimediaEvents.statsd module with ext.wikimediaEvents

It's loaded on all pages for all users. No need for this to be a separate module.
The configuration is already evaluated client-side. The extra size is minimal
and considering this extension is Wikimedia specific, the potential save of
when this is not enabled is non-existent.

If we need to temporarily disable it, the client-side condition will stop it.
And the server-side code won't help anyway in that case since the module call
is included on all cached pages.

Also:
* Keeping the schema's together (like it was previously).

This is fully backwards-compatible.

Change-Id: I509508fdda5b49b4d20ced835c1a827495987480