Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaEvents  7decf3862b3ab06c66a5d2a2c5c4c8e54cbe7720
Merge ext.wikimediaEvents.search module with ext.wikimediaEvents

It's unconditionally loaded on all pages for all users (by being
a dependency of 'ext.wikimediaEvents').

Modules are not a class or file registry, these all share the same
purpose and semantic intent and should be loaded together.

This is fully backwards-compatible.

Change-Id: I1affb15a51954d336b391a464730c6b483c3f78a