Updated mediawiki/extensions Project: mediawiki/extensions/Flow  85f3e80b48e786f8a25dac46d6008c44ac1983bb
Clear LinkCache when failing to fetch Flow board

WorkflowLoaderFactory tries to fetch the workflow id
from page content. For that, it’ll go through
WikiPage::loadPageData, where it won’t be fetched from
master, but (in WikiPage::loadFromRow) be added to
LinkCache. This is likely causing invalid results
because we *just* moved that page in this same request.

The LinkCache stuff in OptInController.php was useless:
resetArticleID already clears link cache as well.

Moved a bit of code in onTitleMove: we shouldn’t be
potentially changing $oldTitle (by fetching article id)
if it’s not Flow-related.

Bug: T111830
Change-Id: Ia390d95d6253e86cfc779ccf276a0bf035e5e608