Updated mediawiki/extensions Project: mediawiki/extensions/WikidataPageBanner  69cb84e2da566c2ebe5705c28ecd2206112ad87e
Overriding the heading should be configurable

Adds the global configuration variable $wgWPBEnableHeadingOverride
which defaults to true and when disabled will disallow the replacement of
the default heading in the chosen skin.
Certain projects might not want to allow the overriding of the default
heading in Vector/Minerva.
Reviewer should ensure the new configuration parameter gets documented
shortly after +2
Additional changes:
Remove unnecessary CSS rule leftover from the days when we had 2 headings
Change-Id: I668eefcb933409c3e0d24234b9125ea08c2444df