Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  f49554807e1de6d63b620ac5bfe7e7b11362e9c4
Use WikidataPageBanner instead of MobileFrontend specific code from banners

Dependencies are merged and WikidataPageBanner has feature parity
and design parity with MobileFrontend's banner code.

It also gives you many bug fixes that are blocking T101108
With WikidataPageBanner installed but limited to Minerva it can
safely be installed on all wikis in the cluster and be blacklisted for
all skins over than mobile web. With the origin parameter editors can fix up
positioning issues for autogenerated images with {{PAGEBANNER:|origin=-.4,-.3}}

The future is here.

Bug: T101108
Change-Id: I12b10b1875c451e127ab5e39fd6b47f501502b6e