Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  d335c3e8c84ebaabec7c1c30da3c54560f3e0c71
Use ViewFactory for building statementgroupviews in statementgrouplistview

This change introduces a `ViewFactory` method for building `ListItemAdapter`s
for `statementgroupviews` widgets. This `ListItemAdapter` is passed to the
`statementgrouplistview` so that it can create the widgets without having to
pass them all their dependencies.

This change leaves `listview` as an implementation detail of
`statementgrouplistview`. This implementation detail is publicly represented by
having to pass in a `listview.ListItemAdapter`.

It improves separation between the `statementgrouplistview` and
`statementgroupview`, since the `statementgrouplistview` doesn't have to know
how to construct a `statementgroupview` anymore. It also allows to inject a
different implementation, for example in tests.

This is a second step for T75380.

Change-Id: I29f6e77c11bb205a435e46446d4ed52999a3e463