Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  2419d12a905c4855c66043192511e20cbc68a32e
Dashboard: Infinite scroll for translation and suggestion lists

Commit includes non-trivial refactoring as well.

There are several similar looking code in suggestionlist and translationlist
modules. But these two modules are going to diverge a lot in coming patches.
So avoiding early optimization and postponing till we see how much they are
going to be similar.

Change-Id: I34873546604a2efc57e5ef971f7d423128e0a38b