Updated mediawiki/extensions Project: mediawiki/extensions/MobileApp  b4a9cd0f57a15074fef43162d8e92319deccd2e0
Fix image widening on xxhdpi+ devices

Our existing CSS only widens images to 60% for xxhdpi+ devices such
as the Samsung Galaxy S4, because min-device-width is dependent on
screen density; hence, for the S4, (min-device-width: 768px) was
evaluating as true.

So I've added a rule to make the width and max-width 100% after all
for devices in portrait orientation with min-device-width 768px where
the device pixel ratio >= 3.

Bug: T113434

Change-Id: Icf296e5cf9b13a0518352367e9f55c8583b137fc