Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  2507d0bf1e0df02f1bae69342a78f7182d97fd6a
Removing deprecated modules(7)

Fixes all the calls to deprecated modules to reference the new
module names. And if the old or deprecated modules is called or
required in an extension, it will be corrected to reference the new name.
The Thanks extension should be updated so that the removal of the deprecated
call should work.

Bug: T113289
Change-Id: Ifce50aecadd9983145f1cdc5e277c65c84cf1b3a