Updated mediawiki/extensions Project: mediawiki/extensions/Echo  e664cb220f2ca2e19f5f72d8da030d8699e4fada
Update seen time only if the popup is actually open

The notifications can be populated from the popup and outside of it.
We should update the seen time only if the request came from the popup
itself, and only if it is still open by the time the api request
has finished.

Bug: T113645
Change-Id: Id91ddabb85fd582be1890ea420d0559e0cdca167