Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  efd5206ebc2cf0ae8d0996f60545a296a9758602
Relax section id validations in mw.cx.getSourceSection and getTranslationSection

These section ids are generated by parsoid. Usually it is prefixed with 'mw'.
But it is not guaranteed. Sometimes templates assign their own ids to sections

This commit removes the validations for prefixes because we need to accept
ids from templates as well.

Bug: T112253
Change-Id: I6875f6445d43039bc191d02295675b09315bf00c