Updated mediawiki/extensions Project: mediawiki/extensions/Echo  e0cd881aeb3477e36e3437912b5be7869f48a7fb
Mark all read button should always start invisible

Whether the button should appear or not is updated when the badge
is updated, based on having unread notifications. But during construction
the button should be hidden.

When we do update the badge, the button should only appear if there
are unread messages and if the current widget is not 'markReadWhenSeen'
to avoid having the button 'flash' while notifications are marked
read.

Bug: T113629
Change-Id: I002db9e3d8990b5a2cd67bf5d120b1c7eb454362