Updated mediawiki/extensions Project: mediawiki/extensions/ParsoidBatchAPI  360ca5a17734704acfbcbcb7a28c0c01f21b1b8c
In imageinfo, check for invalid title

If an image is requested which has an invalid title, pretend it doesn't
exist, by not adding it to the findFiles() batch. This avoids a fatal
error, and the HTML result is not too bad. For the HTML to really be
correct, Parsoid would detect the invalid title prior to sending the
batch and not even try to construct an image link (this is T113322).

Bug: T114048
Change-Id: I77fd7e8853feeeec89961fe4415d0975270710c4