Updated mediawiki/extensions Project: mediawiki/extensions/Echo  07754e8b718dfd54c8554c9758d606b6fc48a9c4
NotificationBadgeWidget: Fix button icon positioning for change in OOUI

OOUI now uses absolute positioning for these icons, which breaks
the centered text model we'd like to use. Longer-term we should
use a different widget for these buttons, or perhaps a widget of
our own, but for now let's hack around that by undoing the
absolute positioning.

Change-Id: I8cc1fd69eeedde00facec80d82a21d5cc5610992