Updated mediawiki/extensions Project: mediawiki/extensions/GlobalUsage  eecb525ec6e69a479c03a99947021df33c5ba78b
Make hook take array by reference

Locally, array values in hooks don't seem to need the & to be
given by reference, but this doesn't seem to be working on cluster,
and certainly, it should not work.

The added special pages aren't being listed in docs for
list=querypages, and they haven't seem to be updated since
2015-09-07T18:17:41, despite the fact they should be every 3
days.

Change-Id: I4d90c8f862bc8dcce1d87f2383412348c88a2898