Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  1c41b1b84a2eb373995fba685f343fc177d00dff
Speed up a query in fix-stats.php by 99.98%

Seriously speaking, it was scanning the whole of change_tag table for
nothing, because there was no join condition. The table is quite big on
the English Wikipedia, so it took half a minute. There is no need to
check this table in this query at all.

Bug: T112626
Change-Id: Id637ba2076cf4ebed569b15f14110a1adabc93d6