Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  787cedbc0353224f1961319fd08a82de7fb8cba6
Don't mess with LinkCache by using Title in ItemContentTest

Because the Title's id got loaded in the context of the
data provider, the fact that it existed (in the real DB)
got added to the LinkCache. That made subsequent tests
fail, because creating Item:Q2 failed because of the
wrong state of LinkCache.

Change-Id: Ie238317b81939bc70a6ad49a02bec9e5625a04b2