Updated mediawiki/extensions Project: mediawiki/extensions/RestBaseUpdateJobs  c2d6dca537d446c39843d3250c5edbb8970a97e0
RestbaseUpdateJob: Set User-Agent

Requires I5096896d0891c0c, but the option is safely ignored if core
support is not yet merged, so this can be merged at any time.
For T114558 investigation -- we need to identify as many clients as
possible if we're going to characterise anomalous traffic to Parsoid.

Change-Id: Ief29bfa411d1eb3222ccbf23bbf401f392b90c9c