Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  4483cc98de2a62f69447626dead6426c1dab0047
Set suggestions as active tab for new users

A simple check of whether the URL contains campaign parameter
and not having a previous CX source language is done.

Checking if there is an ongoing translation is the perfect way
to determine the active tab. But that is an asynchrounous
api call. Swapping the tabs few seconds after page load does
not look good.

Bug: T112822
Change-Id: I0a4bae9bb03dea2c6f66ccf3eb43e8b4db9e7a64