Updated mediawiki/extensions Project: mediawiki/extensions/Echo  cacccc307a87c3ce563d33382c380c2caf67b066
Follow-up f9e8c5057, 983c7a2463: fix JS error when there are no notifications

We have to account for the placeholder item, which doesn't have a
getModel method. We should ensure the real items and the placeholder
item implement a common interface so we don't need these kinds of
checks all over the place.

Bug: T114853
Change-Id: Iff6f0e85fb7737e09a34e72038b6b41681d95595