Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  820a9ab111b82e4534c355c9060e4c4fbdca9bfa
GlobalCollect: record charged amount, not session amount

If a donor opens a new window with different payment or currency
after we initiate the transaction, our session data will be inaccurate
when they come back from the resultswitcher.

This patch only fixes it for GlobalCollect credit cards - we need
a more general solution.

Bug: T115171
Change-Id: Icd50808607a6c1fcf83f94f0fa449e2d5bcb1e5b