Updated mediawiki/extensions Project: mediawiki/extensions/GoogleLogin  ab606c08d06c926c582191f95e2235bff27012ad
Fix GoogleLogin Login mechanism

While changing from GoogleLoginDB to GoogleUser, there was an issue
with the parameters, GoogleLogin::loginGoogleUser() takes and how
to get these. This results in the problem, that no user was recognized
as "connected with a wiki account" and any user got the CreateOrMerge view.

Follow up: I43f3df473a12

Change-Id: I0166f8673b5bd0bbbf515f7427e8caa97cff67c1