Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  3b1c7dc0defa285d24984101954b3a081a2f5dca
Add strict type checks to EntityParserOutputDataUpdater

An other idea was to split this into two arrays. Problem is: At least
one of the data updaters implements both interfaces. It's still possible
to have two more specific arrays as constructor arguments. I tried that.
But it does not turn out nice. I suggest to stick with what you see in
this patch, please.

Bug: T114220
Change-Id: I8096439daa858d67e6454b7339670403dfde093e