Updated mediawiki/extensions Project: mediawiki/extensions/EventLogging  ad7559f7a053f076e72f56eaee1bfd25ce64f5f7
Use an indicator for code samples, fix no-JS mode

Use the new 1.25 "page status indicators" to place the code samples
popup instead of absolute CSS. Notably, this fixes the position of it in
the MonoBook skin.

Also split the styles into a separate module and load it through
addModuleStyles, so the .client-nojs styles take effect.

The styles are also added to the ParserOutput object instead of in a
hook.

Finally, this also fixes a bug in RTL where the code sample headings
were using the content language direction instead of the user language.

Bug: T62233
Change-Id: I44333d0a0943ab4c92182552acb82c5d64be2f0b