Updated mediawiki/extensions Project: mediawiki/extensions/Flow  7460529fc9a273f21fb198bf362bdf3d5537589b
Style categories the same in nojs and js Flow

Rewrite the output of handlebars' categories to be styled the same
as the mediawiki category design, and adjust the JS rendering to
follow the same styling as well.

Added 'specialCategoryLink' property to mw.flow.data to store the
localized link to Special:Categories page, since there doesn't
seem to be a way to get a properly localized version of that page
in JavaScript.

Bug: T115411
Change-Id: Ie42d32b98ac65ba4a6de9971d26542e9d82addf2