Updated mediawiki/extensions Project: mediawiki/extensions/TimedMediaHandler  372d7ae86edca8f825b115aba5f33a565459e65b
Add a config switch for Theora 2-pass encoding

Patching back and forth as we find bugs is too painful for testing.
Defaulting to theora 2-pass encoding *off* (the safer option), but
now much easier to turn it on to test.

To enable, add to LocalSettings:

  $wgTmhTheoraTwoPassEncoding = true;

and reset transcodes at will. This will produce better-quality
.ogv transcodes at the cost of additional encoding time and opening
you up to exciting bugs in libtheora.

Bug: T115881
Change-Id: I8165408f6fe96df6e72c6a52f5f410565c28d4bb