Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaEvents  0b963011cec65231eb738d84675eae6056c44053
Increase subtest sampling from 1:1000 to 1:200

This brings all users of the satisfaction schema into
the subtest. Because half of these users are in the
control group, and many of those in the test group
don't end up submitting a query that is effected (and
we log it there was the possibliity of an effect), this
should have minimal to no change on the main
satisfaction metric.

Change-Id: I59246cb03ebf66c2c43afab4e6bd31988ed88ecc