Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaEvents  d72dc255ea16acae10df657dd661fe020e72e1d3
Add a change tag to log and RC entry of cross-wiki uploads, too

I incorrectly thought that the RC entry for an upload is associated
with the page revision, but it's actually associated with the log
entry. This meant that our tags displayed in page histories, but
didn't return any results when querying on Special:RecentChanges.

The LocalFile object we're using should know both the rev_id and the
log_id, but we can't get them from it since they're both discarded a
few functions deep in some very nasty code.

Bug: T115328
Change-Id: Ic9c521cc5f69c638f40e8fb03ae147ea0b4af2e0