Updated mediawiki/extensions Project: mediawiki/extensions/Newsletter  523ffc272eab07e20db55c53f2e7aec21a164717
Split ext.newsletter module into a separate CSS module and a JS module

The CSS which colors the subscribe/unsubscribe link is not loaded when
JavaScript is disabled but we should colorize them even when JavaScript is
disabled so split these into two modules. This also allows us to load the
JavaSript module at the bottom.

Change-Id: I2a024ec91eed474c87dc0181294c621b54b464a4