Updated mediawiki/extensions Project: mediawiki/extensions/SmiteSpam  8eb8d80e02bf1fd09d9164524ed59bcbe1aa6e72
Add special page links to Extension:AdminLinks interface

Extension:AdminLinks defines a special page, "Special:AdminLinks", that
holds links meant to be helpful for wiki administrators. Since SmiteSpam
is meant to be helpful for wiki administrators, the links to
Special:SmiteSpam and Special:SmiteSpamTrustedUsers should show up in
the interface.

Bug: T116405
Change-Id: I017519dc5e00424ef135cbe3dcf63ea4dd7fef1a