Updated mediawiki/extensions Project: mediawiki/extensions/TimedMediaHandler  468bdea78dfe2693c309414dad4cb09b6b1c1863
Add VideoJS as an alternative player

This adds a preliminary version of a video.js html5 player to TMH.

The player can be activated by setting wgTmhWebPlayer to "videojs".
The default for this value is "mwembed". This switch is required to
further separate the player specific code until eventually we can
use proper extensions and hooks so that hopefully people can
choose between players.

State:
* This changes the HTML and might wreck havoc on cached WMF content
** This different HTML structure is not definitive
* Support for subtitles is missing.
** WebVTT support is in a separate patch
** We will move subtitle serving to an api request (separate patch)
* No credits yet
* No ogvjs plugin yet
* Support for narrow players is not yet finished
* Support for serving iframes is missing
* Support for time offsets is missing
* Support for enlarging thumbed videos is not implemented

Updating:
Use `grunt update-videojs` to update the videojs stack

Bug: T100106
Change-Id: I0a0b7a1727a3431fb6b924ee379b4cc1226f6547