Updated mediawiki/extensions Project: mediawiki/extensions/DismissableSiteNotice  4479c9f309c0baec44b1955038fdab76968bbfa4
Make dismissable SiteNotices hidden by default so they don't flash

Note this does not affect the accessibility of the SiteNotices.

Besides T108513, this also fixes the dismiss button jumping from the
top of the notice to the side during page load.

Also, updating a bug number in a comment to the Phabricator ID.

Bug: T108513
Change-Id: I7e5a4bedc30e3bfa32995ad38eaef729cca9c362