Updated mediawiki/extensions Project: mediawiki/extensions/Echo  74b50cad30afd71502599fe9d8b0b4a60245416a
Move some boilerplate code into EchoEventFormatter

The conversion of EchoEvent into a EchoEventPresentationModel is now
done by EchoFlyoutFormatter, instead of having each subclass do it. It
also does the canRender() check so subclasses don't need to worry about
it.

The subclasses no longer have access to the underlying EchoEvent object,
so the timestamp is exposed in EchoEventPresentationModel.

Change-Id: I7f0a650373eebac7aa2231b1795b51a6d031ad67