Updated mediawiki/extensions Project: mediawiki/extensions/TrustedXFF  731de54c3be4d8f7d4a9db5a1200e09fcdc30220
Avoid a needless and racy file_exists() check

Instead of calling file_exists() and then trying to open the CDB file, just go
ahead and try to open it and see if it suceeds. This is both more efficient
(because it avoids the file_exists()) and more reliable (because it is not
subject to the race condition whereby the file is removed between the
file_exists() and the open calls). If the file does not exist, allow the
exception from Cdb\Reader to buuble up.

EAFP :) (http://stackoverflow.com/a/11360880)

Change-Id: I095ce9ccf2a2d755e5eeb0ffe4bb94481f3b0c0b