Updated mediawiki/extensions Project: mediawiki/extensions/MassMessage  581beff268d5a50cfff13787d54a255f85aa3b9d
Show an error to the user if the spamlist has no targets on it

MassMessage::getParserFunctionTargets is now cached so it can
be called multiple times.

Got rid of 'page-list' terminology, using 'list of pages' instead.

An API test also had to be updated, since it relied on sending
a message to a page with 0 targets on it.

Bug: T55970
Change-Id: If406411c76abbe5847fa554e93ea2557a73e682b