Updated mediawiki/extensions Project: mediawiki/extensions/VipsScaler  748d6f5db1fb4ada1b561a6fb64643d8598f98f6
Use vips shrink for multi-page files

vips shrink reads the file sequentially, which makes it more
memory efficient on large files. Previously we were not using
it on multipage tiff files beyond the first page, as I didn't
know that the options format for that command is different than
vips im_shrink.

Bug: T117349
Change-Id: Ibcdbda52fd3157272305f05f3db8402f59b73925