Updated mediawiki/extensions Project: mediawiki/extensions/Echo  93387806c2674bb678d970a7748f9c225150830b
Defer onPersonalUrls() DB writes to post-send

We calculate how many messages and alerts are being marked as read, and
subtract them from the count since the database and caches won't be
updated until the end of the request.

For performance, we also get the event_type while doing the
EchoTargetPage lookup query to avoid having to query it individually
later on.

Bug: T117531
Change-Id: I0d9302adf1b4b07a4ff26a04b00d4498aa3fe7ee