Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  bba8dc56edc6248ef87603852be3895de148e0b6
Report ES doc id in RunSearch rather than the title page id

On our test instances when we import an elasticsearch dump we
typically only load ES, and not the mysql database. Due to this
all titles report an id of 0. Instead report the doc id from
elasticsearch. Within CirrusSearch the doc id of a page is always
it's page id.

Change-Id: Ia7d2b92a1720cc5fedab4ba0bc2dd9875c9aba4a