Updated mediawiki/extensions Project: mediawiki/extensions/SemanticForms  bb9be02c92f331be81c667fec8348d4619011ec6
Stop duplicating header of empty section as contents of ultimate section

When using {{{section}}} tags on a form, and editing a page
with such form, and the last section is empty, then on the
form the contents of the last section input would have the
header itself. This no longer happens due to workaround that
ensures there is a trailing new-line that the parsing code
expects.

Bug: T72202
Change-Id: I9347cc14a4f287b87c3bd8d92177648a13f06203