Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  4baef9caca3bf75f56b8fb18280e79f76212d95c
Split SimpleValueRdfBuilder and ComplexValueRdfBuilder

This replaces the monolithic implementation of the RDF mapping
for data values with individual classes for each data type.

NOTE: SimpleValueRdfBuilderTest and ComplexValueRdfBuilderTest
are kept to make sure that all previous test cases still pass.
They should now be redundant though, and can be removed in
a future patch.

Bug: T118500
Change-Id: I6d9b99d657e3dcd1de9072ba5582c3f05a99523b