Updated mediawiki/extensions Project: mediawiki/extensions/Flow  ae477cbbfd65d214b3d6e51069befad2d41f0ba6
Remove wgFlowDefaultWorkflow

Originally, there was a plan to have many different kinds of
workflow_types. We only ever got around to implementing
'discussion' and 'topic'. Other workflows are likely going
to be a separate project entirely.

It no longer makes sense to keep this global around, pretending
these workflow_type thingies are going to be flexible,
configurable.
Flow at this point relies on having this particular structure
in place: there should always be a top-level 'discussion'
workflow, with 'topic' workflows related to it. There is
nothing configurable about that 'discussion' workflow_type,
Flow just needs it...

Change-Id: Ic0040a84eb81ec7abc5a06fe0319b5dcc76aaaf6