Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  8aa22d56e64c540023ee7ace3142e7b5eaa9a039
Fix SpecialListPropertiesTest for non-English wikis

If a language fallback (to English) is applied, which
happens in case the Wiki's language is not English, we
will also show a language name along with it. That makes
the assertion that LanguageNameLookup::getName is never
called fail.

Follows-up: 3e46f26aeb00a4023667257158aaeda2e22e9d7a

Change-Id: I696b90dfd74c59ddba0a994277fd99a52c8d4341