Updated mediawiki/extensions Project: mediawiki/extensions/PagedTiffHandler  cf844949d338583ff38c9b512a61a196ae3c1cd0
Convert hex value to decimal. Don't just cast to int.

Tiffinfo usually returns something like "EXIFIFDOffset: 0x1ab38078".
This code tried to do (int)"0x1ab38078", which is equal to 0,
not 447971448 as is needed.

As a result, PagedTiffHandler was counting exif segments as new pages
where it should not have been.

Bug: T87318
Change-Id: I5d4860f093f9710f76c7c9e7af50f60e85fe6715