Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  f48f49b1f865e1ba811994057da33de21533a0d7
Introduce SnakUrlExpander

SnakUrlExpander takes a snak and returns a URL (or URI).

PropertyInfoSnakUrlExpander looks up a URL pattern
in a PropertyInfoStore and fills in a string value
from the snak to generate the URL.

SnakUrlExpander is intended to be used by formatters to generate
links for external IDs, and by RDF mappigns to generate URIs for
external IDs. It may also be used to generate derived values
representing these URLs and URIs for serialization.

Bug: T112762
Change-Id: Ib5e0ec2188731a6c648a806d0f7717e47084a50a