Updated mediawiki/extensions Project: mediawiki/extensions/GoogleLogin  e24884e52e0a17ece85c94930c85c85f1490837a
Implement support for UserMerge

Merge and delete is supported, where a merge will add the google connection
of the old account to the new one, if the new one is not Anonymous and does not
already has a google connection. Otherwise the connection will be left as is,
or deleted (for a merge with Anonymous).

Bug: T114897
Change-Id: Id67ca54e0b3c4170a8395801e91738227e9ef24b