Updated mediawiki/extensions Project: mediawiki/extensions/HitCounters  b3634d33c7519b5d679355ffe6a151b7611a0e0e
Redefine HitCounters magic words

{{NUMBEROFVIEWS}} is now the total number
of visits to the site, as it was before
this extensions got extracted from
the MediaWiki core.

{{NUMBEROFPAGEVIEWS}} is a new magic word
that provides a number of visits to the page,
as presented in the page's footer.
This is what {{NUMBEROFVIEWS}} reported
in the 0.1 and 0.2 versions of this extension.

Bump extension version to 0.3

Bug: T105862
Change-Id: Ie2ce867643a465b58272a09bf3abedc3b5947e34