Updated mediawiki/extensions Project: mediawiki/extensions/Newsletter  ddd5f77c98f57e897d167852131613febcabd524
Cleanup notification sent for new issues

NOTE: This is a BREAKING CHANGE. All previous notifications will
have to be purged from the table as they are not supported. This is
not backwards-compatible as the extension is still in beta status and
I don't want to keep code just for supporting those old notifications
(probably only in development instances).

* Escape raw HTML
* Add secondary link
* Add payload ("summary")
* Set category so that prefs are actually used
* Add agent info
* Expand i18n messages documentation
* Set default user options (web and email) for newsletter-announce notifs
* Remove redundant code; just use code provided by Echo

Bug: T116382
Change-Id: Iaca3a3d88b08adf74333aeaf6b328fc35fe65b97