Updated mediawiki/extensions Project: mediawiki/extensions/Graph  4cb4da7946f54998109284ae7eab6dcb4bc47c11
Implement vega 2.0 support

By default, use the $wgGraphDefaultVegaVer version,
but if the default is 1 and any of the graphs on a page
contain "version" tag higher than 1, use vega2.

Note: js/graph1.js and js/graph2.js are almost identical on purpose:
the graph1 should go away as soon as we migrate to Vega 2.0.

the d3.layout.cloud.js was removed as it wasn't being used at the moment.

Bug: T106103
Change-Id: I562080995111af8aee652842e360fe456326150e