Updated mediawiki/extensions Project: mediawiki/extensions/CharInsert  fe648d41005de73556401978efe53ff351725a48
Remove inline event handler js from charinsert

The primary goal is compatability with I80f6f469ba. However,
I think getting rid of inline event handlers is a good idea
regardless.

This also fixes a bug about edittools on Special:Upload

There are some gadgets that this change is incompatible
with (In particular, if they adjust the innerHTML of an
element containing a charinsert it will cause the browser
to lose the event handler). I'm working on reaching out
to communitites to change the gadget. See T129524.

Note, that this change was originally submitted
as b2f7c4c507e, but was reverted in 28e0ff213d.

This also makes the charinsert links behave less like
links if you have js disabled.

Bug: T11424
Bug: T92781
Bug: T88922
Change-Id: I2d298040ca330293ec9673e5dcc1f4c40b48e752