Updated mediawiki/extensions Project: mediawiki/extensions/MultimediaViewer  6dddb2f7dbf5f6ce26135457fdb2dbce2c897580
Do not set up the overlay on irrelevant hash changes

On not-MMV -> not-MMV hash changes the bootstrap class has set
up and immediately torn down the overlay. Besides slowing things
down, this broke TOC navigation on Chrome in an ugly way.

Add a special case to skip the overlay loading when the new hash
does not contain #/media. This is a quick-and-dirty fix; the whole
loading and hash handling could use a rewrite.

Bug: T119854
Change-Id: I5494903dfe778e533773ff142fdf756475c2df32