Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaEvents  19920f703e700b0deb5522c83dc218124be98a10
Only report timing.editResponseTime if start/end measures are present

On the job runners, the requestShutdown measure may not be set when the
deferred callable that reports editResponseTime is called, causing the value we
report to be bogus. As of I75af5273, Timing::measure() will return false in
such cases.

Change-Id: Ia0774e57ec4625bd2579529e8812b259381472ae