Updated mediawiki/extensions Project: mediawiki/extensions/EventBus  57d9c53ebe004155196a125e838a46d15750a6af
handle parent_revision_id on page creation

When a new page is created, the hook is called with a `$baseRevId` value of
FALSE (to indicate there is none).  In such cases, we should omit the
`parent_revision_id` attribute of events entirely, as it is a) not required,
but b) must have a value greater than 1 if supplied.

Bug: T116786
Change-Id: I65d699c7f6d141cb7b890d5e9ed567fe27843a80