Updated mediawiki/extensions Project: mediawiki/extensions/BlueSpiceFoundation  05b3ae708e8088693d7f1912c4cad1ce0ec14c5d
Show BlueSpice Version in the appropriate column

In Special:Version, the version number was merged with the extension
name. Now it shows up in the version column. Also removed "MediaWiki"
from the display name as this is obvious in a MediaWiki installation.

Patch Set 2: Unified URLs according to Kghbln's comments

Change-Id: I4af0ba8cd58dfd10481487d9cea252d4f23b7423