Updated mediawiki/extensions Project: mediawiki/extensions/SyntaxHighlight_GeSHi  3203eaf554054a30770cca21dfce80e9a115ed6b
Fix inheritance of ve.ui.MWSyntaxHighlightDialogTool

From the description of the commit message[1] for WindowTool (which replaces
DialogTool) and FragmentWindowTool, the
ve.ui.MWSyntaxHighlightDialogTool should be a FragmentWindowTool,
and not a WindowTool/DialogTool.

In fact, the bug is caused by the missing isCompatibleWith method
in WindowTool.

[1] I3a0e761f0d6e942d503ec

Follow up: I676af9c50b09ef007926240eb2327d1a0fdd80f5

Bug: T121596
Change-Id: I0fa440d845c1dea34b623de0a36fb57cebf60f78