Updated mediawiki/extensions Project: mediawiki/extensions/CentralNotice  e55345fc20a90352c4dbea05df40f62fabbd3283
ext.centralNotice.geoIP: Document situation around cookies

The previous comment was somewhat confusing.

I intially thought the comment was wrong and that the cookie-update works fine.
Because the 'en.wikipedia.org' cookie would override (shadow) the
'.wikipedia.org' cookie and things should be fine.

But alas, cookies don't work that way. There is no concept of shadowing.
The 'document.cookies' interface contains them all and it's up to the client
to decide which one to use. Which is difficult since the interface doesn't
expose any meta data about the cookies (their path or their domain etc.).
It ends up like this:

> document.cookies = "GeoIP=:::::v6; * GeoIP=US:CA:*:v4;"

Change-Id: I645255803fe1c011901b4be2c24da20dcbeed91a