Updated mediawiki/extensions Project: mediawiki/extensions/EventLogging  de2320d4e61c4a934146e3cf7ef656e4ccf1a921
Allow client-side schemas to be registered in extension.json

Use the "attributes" feature of extension registration to allow
extensions to set "EventLoggingSchemas" as a top-level key in their
extension.json file.

The new precedence order is:
* 'EventLoggingSchemas' hook
* 'EventLoggingSchemas' attribute in extension.json
* $wgEventLoggingSchemas global

Also clean up the function documentation to use the normal doxygen tags
and add one for the new extension.json style.

Change-Id: I01aa2887a9b98c2ef52cacc3a892390a72b4a3ba