Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  e69f1d15bd47439d29b135569c0a87dcda491732
Allow other skins to be the mobile skin

This also allows a hidden user preference `mobileskin` that allows users to
override their mobile skin.

In this mode, a skin can be loaded with HTML formatted to support
section collapsing HTML and target mode mobile.
e.g. ?useskin=vector&useformat=mobile

Exposure of this is out of scope for this patch and can be explored
in a later patchset

Also the ability to configure as a user preference should be done in
a follow up patch.

See I7b796385ce9786c2c7900e5acc36b1b4dbbb5b97
Change-Id: I7c000165c61a4bd2094f23b6ffd96335f7f7818f