Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  efd76ac77a23f9229d9413b02fb876b5dcbb55ca
Do not re-render nearby items on middle click

Clicking on a nearby item changes the hash so that when the
user comes back to the Special:Nearby page the browser can
scroll to the last clicked item position. On hash change
though the nearby items are re-rendered. The patch allows
changing the hash only when a Alt, Ctrl, Shift, or Meta
keys are not pressed or the middle mouse button is not clicked.

Bug: T114389
Change-Id: I3768abf8d3c94dcabfd0211683dcc896910b7d63