Updated mediawiki/extensions Project: mediawiki/extensions/NewSignupPage  89b74e7bf2e874e37c7d97dda682097db65d1b21
Version 0.8: removed the username existence checker in favor of core MW code

At least MW 1.25 (and obviously newer versions) has a built-in thing for
checking whether the supplied username is or isn't in use, and it appears
to actually work as intended, unlike NewSignupPage's.
Furthermore this simplifies the codebase a bit and slims down the amount
of JS loaded on Special:UserLogin/signup.

Password mismatch checker is not (yet?) in core, so that feature remains
here.

Also swapped "jQuery" to "$" in the JS file, because why not?

Change-Id: I3ce2e96e74656da7f744e19c38d7c9e4545656ee