Updated mediawiki/extensions Project: mediawiki/extensions/OpenBadges  3fcd1e774939c60afe70ccd03fa48fa50f48a602
Issue badges through API and add e-mail requirments

Adds a new API module which allows badges to be issued. This also
meant refactoring the existing API module and moving some functions
to an abstract API module.

Adds two config variables which control the e-mail requirments for being
issued a badge, through either API or Special page, and for reetrieving
a badge assertion (as e-mail could be changed in between).

In passing thumb width for the badge image, when requested through the api,
was changed into a config variable (T103845).

Bug: T103845
Bug: T103827
Change-Id: Id63829bac66dfc0a836af932a526c1bb616aac0c