Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  4212f2700a50c5c3e4cc897290b60ecbb6527272
Fix lead section handling in MobileFrontend

Page#getLeadSectionElement always assumes, that the lead section is wrapped in the
first div element under the .content class div element. This, in fact, isn't true,
it depends on more settings, because the lead section is makred within the process
of marking the expandable sections (which isn't done on special pages, category pages
and so on). In fact, getLeadSectionElement never worked reliable and the worst case
scenario would be "don't trust the value". This especailly makes problems for the editor,
which uses this information to determine where to link the top ca-edit button to. This
could result in the problem, that the top/lead section can't be edited, if the first div
element is empty (because it's not the "real" lead section (because it's not marked)).

This change rewrites the code a little bit, which handles the marking of the lead section
and the expandable section headings. Now, the lead section is always marked, no matter,
if expandable sections are enabled on this page or not. Furthermore, instead of relying
on the amount of div containers (which is very fuzzy), Page#getLeadSectionElement now uses
a newly added class for the lead section (mw-mobilefrontend-leadsection) to always return
the correct lead section element.

Bug: T122471
Change-Id: I310822cffd43115d56c1b253e564f79ca0acdf79