Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  7aa52f6d556e97d5329a772c6b5aaa52eb3a82c0
Require IContextSource as constructor param in EditFilterHookRunner

In places that I see, this always is provided already to the
constructor. If context is not instanceof MutableContext,
then create a DerivativeContext for it.

Context is also always provided for WikibaseRepo::getApiHelperFactory,
so we can remove the option of that being null there.

WikibaseRepo::newEditEntityFactory sometimes has $context as
null (from UpdateRepoOnDelete etc.) so we can allow that,
and then default to RequestContext::getMain there.

Change-Id: I3e82e8ff31d89930313c59a366214ea9faa2df53