Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  6be2735a3788b129fd55433e5f2300455a3af1eb
Only delete cookies which are actually set

Some API clients seem to be confused by cookie deletion.
Prevent cookie deletion on the first leg of the API login sequence
(for a client with an empty cookie jar) by only emitting deletion
headers for cookies which are set in the current request.

Bug: T124252
Change-Id: Iab141f39ba2a24d730dc4dbda286e02a24a2ab4f
Depends-On: I180e094ea32f951e22adab2ec87d16e5de7cef97