Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  328b5800a1a212249c892c3c2e23ec5211618012
Can't use getPersistedSessionId any more

*sigh* another core change, another bit of breakage. There doesn't
seem to be a good way to just see if a session exists with
SessionManager. getSessionInfoForRequest would tell us, but it's
private. So this commit does away with session_exists, which was
barely used anyway, and just lets you ensure that a session does
exist.

Change-Id: Ic19b0562fc3aa5cd91eacdcb3da0145990ebb72f